+++ v2/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Disable AF mode in tests (not supported since 3.12)
 - Allow skipping validation for Database and Collection existence
 - Add support for Graph API
+- Add support for Graph API - Vertex
 
 ## [2.0.3](https://github.com/arangodb/go-driver/tree/v2.0.3) (2023-10-31)
 - Add optional status code checks. Consistent return of response